modules/finance/requirements.txt
# OCR & Image Processing
pytesseract>=0.3.10
Pillow>=10.0.0

# Fuzzy Matching
thefuzz[speedup]>=0.22.0
rapidfuzz>=3.0.0

# AI / LLM
openai>=1.0.0  # Common interface for Ollama/DeepSeek

# Shared
# obsidian-brain-shared @ file:///shared

# Utils
python-dotenv>=1.0.0
watchdog>=4.0.0
